fix: Move call to GitHub integration tasks out from trigger_feature_state_change_webhooks
